requirements.txt
flask
python-dotenv
diffusers
transformers
accelerate
safetensors
peft
pillow
requests
runwayml
openai
torch
torchvision
torchaudio
timm
scikit-learn
pandas
opencv-python
opencv-python-headless
numpy
matplotlib